Update Cafeteria Manager: Linux release files and download options
Explanation: This commit updates the `Cafeteria Manager` project's release files for Linux platforms, including .deb and .AppImage, as well as adding a new check to determine whether to offer the .deb option (index 2) when using Linux. The existing download options are also adjusted accordingly based on the updated filenames. Additionally, minor changes are made to the `DownloadButton` component in order to accommodate these updates.